requirements.txt
Flask
netmiko
psycopg2-binary
python-dotenv